create new space_view_bar_chart
