+++ Cargo.toml
@@ -45,9 +45,10 @@ re_sdk = { path = "crates/re_sdk", version = "0.7.0-alpha.0", default-features =
 re_sdk_comms = { path = "crates/re_sdk_comms", version = "0.7.0-alpha.0", default-features = false }
 re_smart_channel = { path = "crates/re_smart_channel", version = "0.7.0-alpha.0", default-features = false }
 re_space_view = { path = "crates/re_space_view", version = "0.7.0-alpha.0", default-features = false }
+re_space_view_bar_chart = { path = "crates/re_space_view_bar_chart", version = "0.7.0-alpha.0", default-features = false }
+re_space_view_spatial = { path = "crates/re_space_view_spatial", version = "0.7.0-alpha.0", default-features = false }
 re_space_view_text = { path = "crates/re_space_view_text", version = "0.7.0-alpha.0", default-features = false }
 re_space_view_text_box = { path = "crates/re_space_view_text_box", version = "0.7.0-alpha.0", default-features = false }
-re_space_view_spatial = { path = "crates/re_space_view_spatial", version = "0.7.0-alpha.0", default-features = false }
 re_string_interner = { path = "crates/re_string_interner", version = "0.7.0-alpha.0", default-features = false }
 re_tensor_ops = { path = "crates/re_tensor_ops", version = "0.7.0-alpha.0", default-features = false }
 re_time_panel = { path = "crates/re_time_panel", version = "=0.7.0-alpha.0", default-features = false }
