+++ crates/re_space_view_bar_chart/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+authors.workspace = true
+description = "A Space View that shows a single bar chart."
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+name = "re_space_view_bar_chart"
+publish = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+include = ["../../LICENSE-APACHE", "../../LICENSE-MIT", "**/*.rs", "Cargo.toml"]
+
+[package.metadata.docs.rs]
+all-features = true
+
+[dependencies]
+re_arrow_store.workspace = true
+re_components.workspace = true
+re_data_store.workspace = true
+re_data_ui.workspace = true
+re_log_types.workspace = true
+re_log.workspace = true
+re_query.workspace = true
+re_renderer.workspace = true
+re_space_view.workspace = true
+re_tracing.workspace = true
+re_ui.workspace = true
+re_viewer_context.workspace = true
+
+egui.workspace = true
+egui_extras.workspace = true
+vec1.workspace = true
