+++ crates/re_space_view_bar_chart/src/lib.rs
@@ -1,0 +1,8 @@
+//! Rerun Text Space Views
+//!
+//! A Space View that shows text entries in a table and scrolls with the active time.
+
+mod scene_part;
+mod space_view_class;
+
+pub use space_view_class::BarChartSpaceView;
