+++ crates/re_viewer/Cargo.toml
@@ -51,6 +51,7 @@ re_log.workspace = true
 re_memory.workspace = true
 re_renderer = { workspace = true, default-features = false }
 re_smart_channel.workspace = true
+re_space_view_bar_chart.workspace = true
 re_space_view_spatial.workspace = true
 re_space_view_text_box.workspace = true
 re_space_view_text.workspace = true
