+++ crates/re_viewer/src/app.rs
@@ -943,6 +943,7 @@ fn populate_space_view_class_registry_with_builtin(
     space_view_class_registry.add::<re_space_view_text::TextSpaceView>()?;
     space_view_class_registry.add::<re_space_view_text_box::TextBoxSpaceView>()?;
     space_view_class_registry.add::<re_space_view_spatial::SpatialSpaceView>()?;
+    space_view_class_registry.add::<re_space_view_bar_chart::BarChartSpaceView>()?;
     Ok(())
 }
