--- crates/re_viewport/src/view_bar_chart/mod.rs
@@ -1,5 +1,0 @@
-mod scene;
-pub(crate) use self::scene::SceneBarChart;
-
-mod ui;
-pub(crate) use self::ui::{help_text, view_bar_chart, BarChartState};
